+++ test/unit/fuzz-tools.test.js
@@ -445,6 +445,7 @@ import * as letournelAcetabulumV466 from '../../lib/letournel-acetabulum-v466.js
 import * as bromageScaleV467 from '../../lib/bromage-scale-v467.js';
 import * as brouetCryoglobulinemiaV468 from '../../lib/brouet-cryoglobulinemia-v468.js';
 import * as steinbrockerRaV469 from '../../lib/steinbrocker-ra-v469.js';
+import * as larsenRaV470 from '../../lib/larsen-ra-v470.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -835,6 +836,7 @@ const MODULES = {
   'bromage-scale-v467.js': bromageScaleV467,
   'brouet-cryoglobulinemia-v468.js': brouetCryoglobulinemiaV468,
   'steinbrocker-ra-v469.js': steinbrockerRaV469,
+  'larsen-ra-v470.js': larsenRaV470,
   'tb-testing.js': tbTesting,
 };